Implements auth context
This auth context checks if there's a jwt in the localStorage and if it's valid, and can also refresh it in case it's valid but expired. The pages that need any info about the session state or the current user will be able to make use of it throught the useAuth hook.