implements login form
since the entire page will be protected, i need a login form where users can be redirected to.
Most of the logic is taken from the form in the main blog, so it already has error handling on client and server. The useAuth hook was implemented so it won't load until the token is resolved.