Creates PopupMenu component
this popup menu will receive its children and a function to dismount itself when clicked outside of it (the parent component has to dismount it since it's the one rendering it).
The children, wich are supposed to be grid elements with an icon and a label, are displayed in a list.

This is being added as a replacement of the action buttons on cards